+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.27.1] - 2024-10-04
+
+### Fixed
+
+- Properly offload datapoint parsing to a background thread during `load` in the async client.
+- Progress callbacks during `load` are now always done from the main thread in the sync client.
+  This resolves an issue when using it in `streamlit` to update a progress bar, since interacting with
+  streamlit widgets from a background thread is not allowed without setting some streamlit context in the background
+  thread.
+- Accept any return type for show_progress callbacks in `load`
+
 ## [0.27.0] - 2024-10-02
 
 ### Changed
@@ -41,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.27.0...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.27.1...HEAD
+[0.27.1]: https://github.com/tilebox/tilebox-python/compare/v0.27.0...v0.27.1
 [0.27.0]: https://github.com/tilebox/tilebox-python/compare/v0.26.0...v0.27.0
 [0.26.0]: https://github.com/tilebox/tilebox-python/compare/v0.25.1...v0.26.0
 [0.25.1]: https://github.com/tilebox/tilebox-python/compare/v0.25.0...v0.25.1